# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /**
   * @minLength 1
   * @maxLength 30
   * @pattern ^[a-zA-Z0-9]+$
   */
  username: string;
  /**
   * @minLength 8
   * @pattern ^(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9]).{8,}$
   */
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface ResetPasswordRequest {
  username: string;
  /**
   * @minLength 8
   * @pattern ^(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9]).{8,}$
   */
  newPassword: string;
}

export interface SessionUser {
  id: number;
  username: string;
  isGuest: boolean;
}

export interface AuthResponse {
  user: SessionUser;
  /** JWT for use in Authorization header */
  token: string;
}

export interface NewsArticle {
  id: string;
  title: string;
  description: string;
  url: string;
  source: string;
  publishedAt: string;
  isRelevantToDevTesting: boolean;
  tags: string[];
}

export interface NewsResponse {
  articles: NewsArticle[];
  total: number;
  page: number;
  limit: number;
}

export interface SavePostRequest {
  articleUrl: string;
  articleTitle: string;
  articleSource: string;
  articleDate: string;
  articleDescription?: string;
}

export interface SavedPost {
  id: number;
  userId: number;
  articleUrl: string;
  articleTitle: string;
  articleSource: string;
  articleDate: string;
  articleDescription?: string;
  savedAt: string;
}

export interface SavedPostsResponse {
  posts: SavedPost[];
}

export type GetNewsParams = {
  page?: number;
  limit?: number;
};