# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionUser } from "./sessionUser";

export interface AuthResponse {
  user: SessionUser;
  /** JWT for use in Authorization header */
  token: string;
}