# lib/api-zod/src/generated/types/newsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticle } from "./newsArticle";

export interface NewsResponse {
  articles: NewsArticle[];
  total: number;
  page: number;
  limit: number;
}