# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  /**
   * @minLength 1
   * @maxLength 30
   * @pattern ^[a-zA-Z0-9]+$
   */
  username: string;
  /**
   * @minLength 8
   * @pattern ^(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9]).{8,}$
   */
  password: string;
}