# lib/api-zod/src/generated/types/resetPasswordRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordRequest {
  username: string;
  /**
   * @minLength 8
   * @pattern ^(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9]).{8,}$
   */
  newPassword: string;
}