# lib/api-zod/src/generated/types/savedPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedPost {
  id: number;
  userId: number;
  articleUrl: string;
  articleTitle: string;
  articleSource: string;
  articleDate: string;
  articleDescription?: string;
  savedAt: string;
}