# lib/api-zod/src/generated/types/savedPostsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { SavedPost } from "./savedPost";

export interface SavedPostsResponse {
  posts: SavedPost[];
}