# lib/api-zod/src/generated/types/sessionUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI News Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionUser {
  id: number;
  username: string;
  isGuest: boolean;
}